Refactor distribution code to one file
