+++ d100/ast/dice.py
@@ -7,8 +7,7 @@ from .die import DiceSize, Die
 from .node import ASTNode, Number
 from .operators import AdvantageCategory, Operator, OperatorCategory, Selector
 from ..context import RollContext
-from ..distribution import Distribution
-from ..distribution.calculate import ConvolutionDistributionBuilder, DiscreteDistributionBuilder
+from ..distribution import Distribution, ConvolutionDistributionBuilder, DiscreteDistributionBuilder
 from ..errors import RollError, RollValueError
